Move gui imports from console/utils into console/gui_utils
